Trying to share most of the code between RequestDispatcher and SpringViewResolver providers
